% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{data_blocks_heatmap}
\alias{data_blocks_heatmap}
\title{Heatmaps of several data blocks.}
\usage{
data_blocks_heatmap(blocks, show_color_bar = TRUE)
}
\arguments{
\item{blocks}{List. List containing matrices to plot.}

\item{show_color_bar}{Boolean. Whether or not to display the color bars.}
}
\description{
Heatmaps of several data blocks.
}
\examples{
library(ggplot2)
library(cowplot)
blocks <- sample_toy_data()
data_blocks_heatmap(blocks)

}
