% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{data_heatmap}
\alias{data_heatmap}
\title{Heatmap of a single data matrix.}
\usage{
data_heatmap(data, show_color_bar = TRUE, title = "", xlab = "",
  ylab = "")
}
\arguments{
\item{data}{Matrix. The data matrix to plot.}

\item{show_color_bar}{T/F. Whether or not to show the color bar.}

\item{title}{String. Optional plot title}

\item{xlab}{String. Optional x axis label.}

\item{ylab}{String. Optional y axis label.}
}
\description{
Heatmap of a single data matrix.
}
\examples{
data <- sample_toy_data()
Y_obs <- data[[2]][['obs']]
data_heatmap(Y_obs)

}
