% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{decomposition_heatmaps}
\alias{decomposition_heatmaps}
\title{Heatmaps of JIVE decomposition of several blocks.}
\usage{
decomposition_heatmaps(block_decompositions)
}
\arguments{
\item{block_decompositions}{List containing the K JIVE decomposition lists.}
}
\description{
Heatmaps of JIVE decomposition of several blocks.
}
\examples{
data <- sample_toy_data()
library(cowplot)
data_blocks_heatmap(data)

}
