% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities_of_interest.R
\name{get_block_full}
\alias{get_block_full}
\title{Returns one of the full JIVE decomposition matrices.}
\usage{
get_block_full(ajive_output, k, type)
}
\arguments{
\item{ajive_output}{Output from the ajive function.}

\item{k}{Integer. Which block.}

\item{type}{String. type is either 'joint', 'individual' or 'noise'.}
}
\value{
A matrix of dimension equal to the dim(block[[k]]).
}
\description{
get_block_full returns one of J, I or E.
}
