% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities_of_interest.R
\name{get_block_loadings}
\alias{get_block_loadings}
\title{AJIVE loadings}
\usage{
get_block_loadings(ajive_output, k, type)
}
\arguments{
\item{ajive_output}{Output from the ajive function.}

\item{k}{Integer. Which block.}

\item{type}{String. type is either 'joint' or 'individual'.}
}
\description{
AJIVE loadings
}
