% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities_of_interest.R
\name{get_block_scores}
\alias{get_block_scores}
\title{Returns the AJIVE block scores.}
\usage{
get_block_scores(ajive_output, k, type, normalized)
}
\arguments{
\item{ajive_output}{Output from the ajive function.}

\item{k}{Integer. Which block.}

\item{type}{String. type is either 'joint' or 'individual'.}

\item{normalized}{Boolean. Whether or not to normalize the scores.}
}
\value{
A matrix of dimension n x rank-(joint/individual-k).
}
\description{
Representation of a data block.
}
\details{
The scores are a lower dimensional representation of a data block (e.g. latent variables).

The scores can either be normalized (i.e. just U) or unnormalized (i.e. UD)/

If the SVD of I_k is U, D, V then U are the unformalized individual scores for block k.
}
