% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities_of_interest.R
\name{get_common_normalized_scores}
\alias{get_common_normalized_scores}
\title{Returns the common normalized scores.}
\usage{
get_common_normalized_scores(ajive_output)
}
\arguments{
\item{ajive_output}{Output from the ajive function.}
}
\value{
A matrix of dimension n x joint-rank that is a basis of the joint space.
}
\description{
Represents the joint signal among the data blocks.
}
