% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities_of_interest.R
\name{get_individual_rank}
\alias{get_individual_rank}
\title{Individual rank of block k.}
\usage{
get_individual_rank(ajive_output, k)
}
\arguments{
\item{ajive_output}{Output from the ajive function.}

\item{k}{Integer. Which block.}
}
\description{
Individual rank of block k.
}
