% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities_of_interest.R
\name{get_joint_rank}
\alias{get_joint_rank}
\title{Joint rank.}
\usage{
get_joint_rank(ajive_output)
}
\arguments{
\item{ajive_output}{Output from the ajive function.}
}
\description{
Joint rank.
}
