% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajive_decomposition.R
\name{get_joint_scores}
\alias{get_joint_scores}
\title{Computes the joint scores.}
\usage{
get_joint_scores(
  blocks,
  block_svd,
  initial_signal_ranks,
  sv_thresholds,
  n_wedin_samples = 100L,
  n_rand_dir_samples = 100L,
  joint_rank = NA
)
}
\arguments{
\item{blocks}{A list of numeric matrices with same number \eqn{N} of rows
specifying the data blocks to be analyzed.}

\item{block_svd}{A list of lists specifying the SVDs of each data block.}

\item{initial_signal_ranks}{An integer vector specifying the initial ranks
for the signal matrices.}

\item{sv_thresholds}{A numeric vector specifying the singular value
thresholds from the initial signal rank estimates.}

\item{n_wedin_samples}{An integer value specifying the number of Wedin bound
samples to draw. Defaults to \code{100L}.}

\item{n_rand_dir_samples}{An integer value specifying the number of random
direction bound samples to draw. Defaults to \code{100L}.}

\item{joint_rank}{An integer valye specifying the rank of the joint space.
Defaults to \code{NA} in which case it is estimated from the data.}
}
\value{
A numeric matrix of size \eqn{N \times r_J} storing the joint scores.
}
\description{
Estimate the joint rank with the wedin bound, compute the signal scores SVD,
double check each joint component.
}
