% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_direction_bound.R
\name{get_random_direction_bound}
\alias{get_random_direction_bound}
\title{Estimate the wedin bound for a data matrix.}
\usage{
get_random_direction_bound(n_obs, dims, num_samples = 1000)
}
\arguments{
\item{num_samples}{Integer. Number of vectors selected for resampling procedure.}

\item{n_obs.}{The number of observations.}

\item{n_features.}{The number of features in each data matrix}
}
\value{
rand_dir_samples
}
\description{
Samples from the random direction bound. Returns on the scale of squared singular value.
}
