% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajive_decomposition.R
\name{get_sv_threshold}
\alias{get_sv_threshold}
\title{The singular value threshold.}
\usage{
get_sv_threshold(singular_values, rank)
}
\arguments{
\item{singular_values}{Numeric. The singular values.}

\item{rank}{Integer. The rank of the approximation.}
}
\description{
Computes the singluar value theshold for the data matrix (half way between the rank and rank + 1 singluar value).
}
