% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_fun.R
\name{get_svd}
\alias{get_svd}
\title{Singluar Value Decomposition.}
\usage{
get_svd(X, rank = NULL)
}
\arguments{
\item{X}{Matrix.}

\item{rank}{Integer. Rank of the desired SVD (optional). If rank==0 returns zeros.}
}
\value{
The SVD of X.
}
\description{
Returns a possibly truncated SVD of a data matrix.
}
\details{
Wraps the svd function. Removes the extra singluar values if a truncated svd is computed.
}
