% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wedin_bound.R
\name{get_wedin_bound_samples}
\alias{get_wedin_bound_samples}
\title{Estimates the Wedin bound for a data matrix with the resampling procedure.}
\usage{
get_wedin_bound_samples(X, SVD, signal_rank, num_samples = 1000)
}
\arguments{
\item{X}{Matrix. The data matrix.}

\item{SVD}{List. The SVD decomposition of X.}

\item{signal_rank}{Integer. The estimated signal rank of X.}

\item{num_samples}{Integer. Number of vectors selected for resampling procedure.}
}
\value{
The wedin bound samples.
}
\description{
Returns \eqn{\min(\max(\| E \tilde{V} \|, \|E^\top \tilde{U}\|) / \sigma_{\min(\widetilde{A}),
1)}} from equation (7).
}
