% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_blocks_grid}
\alias{plot_blocks_grid}
\title{Makes a plot for each block in a hotizontal grid.}
\usage{
plot_blocks_grid(blocks, plot_fun, ...)
}
\arguments{
\item{blocks}{List.}

\item{plot_fun}{Function. A plotting function that takes a sigle data matrix and makes a plot}

\item{...}{Additional arguments to be passed to plot_fun}
}
\description{
Makes a plot for each block in a hotizontal grid.
}
