% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{scree_plot}
\alias{scree_plot}
\title{Scree plot}
\usage{
scree_plot(singular_values, title = "")
}
\arguments{
\item{singular_values}{Numeric. The singluar values.}

\item{title}{String. The title of the plot.}
}
\description{
Scree plot
}
\examples{
X <- matrix(rnorm(n=10*20), nrow=20, ncol=10) + matrix(1, nrow=20, ncol=10)
singluar_values <- svd(X)[['d']]
scree_plot(singluar_values)

}
