% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{scree_plot_blocks}
\alias{scree_plot_blocks}
\title{Shows the scree plot for each data block.}
\usage{
scree_plot_blocks(blocks)
}
\arguments{
\item{blocks}{The data blocks.}
}
\description{
Shows the scree plot for each data block.
}
