% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_fun.R
\name{svd_reconstruction}
\alias{svd_reconstruction}
\title{Reconstruces the original matrix from its SVD.}
\usage{
svd_reconstruction(decomposition)
}
\arguments{
\item{decomposition}{List. List with entries 'u', 'd', and 'v'from the svd function.}
}
\value{
Matrix. The original matrix.
}
\description{
Computes UDV^T to get the approximate (or full) X matrix.
}
