% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wedin_bound.R
\name{wedin_bound_resampling}
\alias{wedin_bound_resampling}
\title{Resampling procedure for the wedin bound}
\usage{
wedin_bound_resampling(X, perp_basis, right_vectors, num_samples = 1000)
}
\arguments{
\item{X}{Matrix. The data matrix.}

\item{perp_basis}{Matrix. Either U_perp or V_perp: the remaining left/right singluar vectors of X after estimating the signal rank.}

\item{right_vectors}{Boolean. Right multiplication or left multiplication.}

\item{num_samples}{Integer. Number of vectors selected for resampling procedure.}
}
\description{
Resampling procedure for the wedin bound
}
