% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajive.R, R/ajive_decomposition.R
\docType{package}
\name{ajive}
\alias{ajive}
\alias{ajive-package}
\title{ajive: Angle based Joint and Individual Variation Explained.}
\usage{
ajive(blocks, initial_signal_ranks, full = TRUE,
  n_wedin_samples = 1000, n_rand_dir_samples = 1000, joint_rank = NA)
}
\arguments{
\item{blocks}{List. A list of the data matrices.}

\item{initial_signal_ranks}{Vector. The initial signal rank estimates.}

\item{full}{Boolean. Whether or not to store the full J, I, E matrices or just their SVDs (set to FALSE to save memory).}

\item{n_wedin_samples}{Integer. Number of wedin bound samples to draw for each data matrix.}

\item{n_rand_dir_samples}{Integer. Number of random direction bound samples to draw.}

\item{joint_rank}{Integer or NA. User specified joint_rank. If NA will be estimated from data.}
}
\value{
The JIVE decomposition.
}
\description{
The ajive package provides functions to do:
foo, bar and baz.

Computes the JIVE decomposition.
}
\section{Foo functions}{

The foo functions ...
}

\examples{
blocks <- sample_toy_data(n=200, dx=100, dy=500)
initial_signal_ranks <- c(2, 2)
jive_decomp <- ajive(blocks, initial_signal_ranks)

joint_scores <- jive_decomp[['joint_scores']]
J_1 <- jive_decomp[[1]][['joint']][['full']]
U_individual_2 <- jive_decomp[[2]][['individual']][['u']]
individual_rank_2 <- jive_decomp[[2]][['individual']][['rank']]

}
