% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{decomposition_heatmaps}
\alias{decomposition_heatmaps}
\title{Heatmaps of JIVE decomposition of several blocks.}
\usage{
decomposition_heatmaps(blocks, jive_decomposition)
}
\arguments{
\item{blocks}{List. The observed data blocks.}

\item{jive_decomposition}{List containing the JIVE decompositions.}
}
\description{
Heatmaps of JIVE decomposition of several blocks.
}
