% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wedin_bound.R
\name{get_wedin_bound_samples}
\alias{get_wedin_bound_samples}
\title{Esimates the wedin bound for a data matrix with the resampling procedure.}
\usage{
get_wedin_bound_samples(X, SVD, signal_rank, num_samples = 1000)
}
\arguments{
\item{X}{Matrix. The data matrix.}

\item{SVD}{List. The SVD decomposition of X.}

\item{signal_rank}{Integer. The estimated signal rank of X.}
}
\value{
The the wedin bound samples.
}
\description{
returns min(max(||E \tilde{V}||, ||E^T \tilde{U}||) / sigma_min(\widetilde{A}), 1) from equation (7)
}
